+++ packages/ispectify_http/example/main.dart
@@ -4,7 +4,7 @@ import 'package:ispectify_http/ispectify_http.dart';
 void main(List<String> args) async {
   final client = InterceptedClient.build(
     interceptors: [
-      ISpectifyHttpLogger(),
+      ISpectHttpInterceptor(),
     ],
   );