+++ packages/ispectify_ws/lib/ispectify_ws.dart
@@ -1,3 +1,3 @@
-export 'src/logger.dart';
+export 'src/interceptor.dart';
 export 'src/models/_models.dart';
 export 'src/settings.dart';